Add support for Windows ARM64 to the cmake files
As x64 dependencies are currently also required for these builds due to
Qt being cross-compiled, an extra check is added, to ensure that an x64
configure has been done prior to trying to configure/build for ARM64.
As such, the deps folder in Windows now has the arch appended to it, so
that both can be present in the same folder.

`qtserialport` is disabled for this platform, as it had issues being
compiled. It is possible to work around it (it is also related to setting
QT_HOST_PATH), but needs a closer look at a later date.

Signed-off-by: Anthony Roberts <anthony.roberts@linaro.org>